feat(2.9): separate templates
