+++ src/presenter/trip-presenter.js
@@ -1,8 +1,8 @@
 import { render } from '../render';
-import PointsView from '../view/points-view';
-import PointEditView from '../view/point-edit-view';
-import PointView from '../view/point-view';
-import SortingView from '../view/sorting-view';
+import PointsView from '../view/points/points-view';
+import PointEditView from '../view/point-edit/point-edit-view';
+import PointView from '../view/point/point-view';
+import SortingView from '../view/sorting/sorting-view';
 
 export default class TripPresenter {
   pointsComponent = new PointsView();
