+++ src/view/cost/cost.tpl.js
@@ -1,0 +1,5 @@
+export const createCostTemplate = () => (
+  `<p class="trip-info__cost">
+    Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+  </p>`
+);
