+++ src/view/filters/filters.tpl.js
@@ -1,6 +1,4 @@
-import { createElement } from '../render.js';
-
-const createFiltersTemplate = () => (
+export const createFiltersTemplate = () => (
   `<form class="trip-filters" action="#" method="get">
     <div class="trip-filters__filter">
       <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
@@ -20,21 +18,3 @@ const createFiltersTemplate = () => (
     <button class="visually-hidden" type="submit">Accept filter</button>
   </form>`
 );
-
-export default class FiltersView {
-  getTemplate() {
-    return createFiltersTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
