+++ src/view/loading/loading-view.js
@@ -1,0 +1,21 @@
+import { createElement } from '../../render';
+
+import { createLoadingTemplate } from './loading.tpl';
+
+export default class LoadingView {
+  getTemplate() {
+    return createLoadingTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
