+++ src/view/loading/loading.tpl.js
@@ -1,0 +1,1 @@
+export const createLoadingTemplate = () => '<p class="trip-events__msg">Loading...</p>';
