+++ src/view/new-point-button/new-point-button.tpl.js
@@ -1,0 +1,1 @@
+export const createNewPointButtonTemplate = () => '<button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>';
