+++ src/view/point-edit/point-edit-view.js
@@ -1,0 +1,21 @@
+import { createElement } from '../../render';
+
+import { createPointEditTemplate } from './point-edit.tpl';
+
+export default class PointEditView {
+  getTemplate() {
+    return createPointEditTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
