+++ src/view/point-edit/point-edit.tpl.js
@@ -1,6 +1,4 @@
-import { createElement } from '../render.js';
-
-const createPointEditTemplate = () => (
+export const createPointEditTemplate = () => (
   `<li class="trip-events__item">
     <form class="event event--edit" action="#" method="post">
       <header class="event__header">
@@ -157,21 +155,3 @@ const createPointEditTemplate = () => (
     </form>
   </li>`
 );
-
-export default class PointEditView {
-  getTemplate() {
-    return createPointEditTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
