+++ src/view/point/point-view.js
@@ -1,0 +1,25 @@
+import { createElement } from '../../render';
+
+import { createPointTemplate } from './point.tpl';
+
+export default class PointView {
+  constructor(point) {
+    this.point = point;
+  }
+
+  getTemplate() {
+    return createPointTemplate(this.point);
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
