+++ src/view/points/points.tpl.js
@@ -1,0 +1,1 @@
+export const createPointsTemplate = () => '<ul class="trip-events__list"></ul>';
