+++ src/view/route/route.tpl.js
@@ -1,0 +1,7 @@
+export const createRouteTemplate = () => (
+  `<div class="trip-info__main">
+    <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
+
+    <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
+  </div>`
+);
