+++ src/view/sorting/sorting.tpl.js
@@ -1,6 +1,4 @@
-import { createElement } from '../render.js';
-
-const createSortingTemplate = () => (
+export const createSortingTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
     <div class="trip-sort__item  trip-sort__item--day">
       <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
@@ -28,21 +26,3 @@ const createSortingTemplate = () => (
     </div>
   </form>`
 );
-
-export default class SortingView {
-  getTemplate() {
-    return createSortingTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
